Fix api project packages
