--- backend/api/src/main/kotlin/com/example/todo/dto/request/TaskCreation.kt
@@ -1,9 +1,0 @@
-package com.example.todo.dto.request
-
-import jakarta.validation.constraints.NotBlank
-
-data class TaskCreation(
-    @field:NotBlank
-    val summary: String,
-    val description: String?,
-)
