--- backend/api/src/main/kotlin/com/example/todo/dto/response/AuthenticatedResponse.kt
@@ -1,8 +1,0 @@
-package com.example.todo.dto.response
-
-/**
- * Access token for auth user.
- */
-data class AuthenticatedResponse(
-    val token: String,
-)
