--- backend/api/src/main/kotlin/com/example/todo/exception/PresentationException.kt
@@ -1,9 +1,0 @@
-package com.example.todo.exception
-
-/**
- * Base exception class for presentation layer.
- */
-open class PresentationException(
-    message: String,
-    cause: Throwable?
-) : RuntimeException(message, cause)
