+++ backend/api/src/main/kotlin/com/example/todo/presentation/dto/request/TaskCreation.kt
@@ -1,0 +1,9 @@
+package com.example.todo.presentation.dto.request
+
+import jakarta.validation.constraints.NotBlank
+
+data class TaskCreation(
+    @field:NotBlank
+    val summary: String,
+    val description: String?,
+)
