+++ backend/api/src/main/kotlin/com/example/todo/presentation/dto/request/TaskEdit.kt
@@ -1,0 +1,12 @@
+package com.example.todo.presentation.dto.request
+
+import com.example.todo.presentation.config.validation.TaskStatus
+import jakarta.validation.constraints.NotBlank
+
+data class TaskEdit(
+    @field:NotBlank
+    val summary: String,
+    val description: String?,
+    @field:TaskStatus
+    val status: String,
+)
