+++ backend/api/src/main/kotlin/com/example/todo/presentation/dto/response/AuthenticatedResponse.kt
@@ -1,0 +1,8 @@
+package com.example.todo.presentation.dto.response
+
+/**
+ * Access token for auth user.
+ */
+data class AuthenticatedResponse(
+    val token: String,
+)
