+++ backend/api/src/main/kotlin/com/example/todo/presentation/exception/PresentationException.kt
@@ -1,0 +1,9 @@
+package com.example.todo.presentation.exception
+
+/**
+ * Base exception class for presentation layer.
+ */
+open class PresentationException(
+    message: String,
+    cause: Throwable?
+) : RuntimeException(message, cause)
